GLSDK: PM-TM: Added a thermal utility
The thermal utility is a shell script that helps user to access
all thermal sysfs entries in a readable manner in one place.
It also enables user to test the temperature emulation feature of the
silicon. To use the thermal emulation user needs to enable
"CONFIG_THERMAL_EMULATION" in kernel menu configuration.
It also displays the available cooling devices and the cooling states.

Signed-off-by: Ravikumar Kattekola <rk@ti.com>